Add a texture
Changes to use a texture rather than specifying colors for the
vertices.

It would be useful to add in the option to render meshes while
only specifying vertex colors (after all, I don't want to do extra
work to texture objects for simple physics demos), but for now
we're just getting textures working.