Add new pipeline for colored meshes
Adds a parallel pipeline to the main render pipeline,
which will handle all meshes that do not have a texture
(which instead only have a single color).

The render pipeline includes bindings for the camera
and the light. It uses the camera, but does not yet use
the light. We include the light binding because we plan
to support blinn-phong shading for solid colored objects
soon.

The screen now displays a floating pentagon and no textured
cube nor light, as we've disabled their draw functions for
testing out this new pipeline.

The ColoredMesh struct is added to the model module.
It's the same as the regular mesh, but includes no texure.
It's meant to be used in the new pipeline.
The actual color data is stored with each vertex in its buffer.
There may be better way to reconcile the Mesh and ColoredMesh
structs, but we'll hold off for now.