Add cube with normals facing inwards
Adds cube with normals facing inwards. We'll use this for
rendering a bouncing ball within a cube, so that we can see
through the cube.

Note (this has been added to TODO.md), that we should change the
cubes in form.rs to use non-averaged vertex normals (i.e. unique
vertices for each face, with 3 overlapping). This would result in
more accurate lighting, since the models are so low poly, which
makes the averaged normal incorrect.