Associate models with instance data
Associates each model/mesh in our scene with corresponding instance
data in the instance buffers.

We use one buffer for static objects in the scene, and one for dynamic
objects. This way, we only need to update the dynamic buffer.

Right now, we just hard code the instance indices to associate them
with models. As our scenes grow more complex, we'd want to abstract this
away. For now, we can keep track of it all ourselves.

Also adds scaling to instance data, and scales up the bounding cube
by a factor of 2. Note that this means for our initial bouncing
ball simulation, we'll just be able to check the center point
of the sphere bouncing off othonormal planes at -1, 1 on each
axis, which simplifies things (since the ball's radius would
extend out to meet the sphere at -2, +2).