Create initial simulation of ball moving
Adds the simulation module, with the submodule bounce.
This will store logic for the simulation of a bouncing ball.

The logic which updates the position of the ball was moved into
its step function.

Further, we improve our handling of time in the game loop according
to https://gafferongames.com/post/fix_your_timestep/
This allows us to separate the display time and the simulation time,
thereby allowing us to simulate the physics at a fixed timestep
(to be made configuratble eventually) while allowing the framerate
to run as fast as possible.
We do add one twist to the code presented there, in that our
step() or integrate() function returns the duration which elapsed
during simulation, and use that to consume the accumulator.
This is to allow for partial time steps in the event of a collision.