Use gravity and air resistance to move ball
Our first physical simulation! We use first order Euler integration
to dictate the movement of the sphere according to the forces acting
on the sphere. We include gravity and air resistance.

There is still no collision handling, so the sphere just phases through
the sphere and falls down forever. We'll add collision handling in
an upcoming change.