Add collisions with planes
Adds collisions with planes.
The collision logic is generic for any orientation, but we don't
show that off yet. Consider doing so in a future change.
The collision response includes friction, and the coefficient
of restitution.

Also fixes a bug where the rendering code was *adding* the position
of the ball to the previous position, rather than simply using the
current position of the sphere. Previously this wasn't noticable
due to the lack of collisions (so it was moving in a straight line
regardless).