Enforce bounding box
As a bit of a hack, check the bounding box and keep the ball
within it. This fixes an issue where the ball would
sometimes clip through the wall in certain scenarios.

The collision handling and resting contact logic should be
improved in the future to obviate the need for this.