Draw basic egui window
Draws a basic egui window over the simulation.
Sets up the gui module to handle storing and rendering the GUI.

To properly decouple rendering logic between our 3D renders and
our GUI, we have the render() function of each return a command
buffer, rather than submitting the queue and presenting the
texture on the surface. We create all the command buffers, and
then submit them at once.

The GUI is just a "hello world" widget for now. We can add
functionality later.