Move lib.rs to bouncing_ball_demo.rs
Moves the code of lib.rs to bouncing_ball_demo.rs. The main fn
will switch on command line arguments to run different _demos,
so we'll choose to not have a central lib.rs for now.

If we make it so that we can switch between the demos without
restarting the app, or otherwise unify them, we can do that
in a lib.rs.