Make Entity struct from Particles
Makes the Particles struct more general to be an Entity,
since it's really just any arbitrary mesh with its instance
data (conceptually - we're still hard coding initialization etc.
But we should be able to generalize it for any mesh/model etc).
We just so happen to be using it for particles now.