Eliminate stack based instance array
Eliminates the use of arrayvec for the instance buffer, which was
an overengineered way to limit the number of particles.
More importantly, an arrayvec is backed by an array on the stack.
This was causing a stack overflow for large amounts of particles.

Now, we can generate many thousands of particles without crashing!