Make graphics module
Makes the graphics module and moves all the graphics-related code
to it.

This isn't a totally perfect boundary right now. I moved GPUInterface
to graphics, e.g., but we might want to use some GPU interface for
e.g. compute shaders, which isn't related to CG per se. You could
also argue that Models/Meshes aren't strictly for graphics, since
they are being used to define geometry for simulations, which don't
necessarily need to be rendered.

All that to say, this may get reorganized in the future. But for
now, I just wanted to get an OK baseline.