Add flocking demo with default scene
Adds the flocking demo boilerplate code with just a default
cube, no simulation yet. Just a scene to build on.

This also adds some generalizations to the Scene implementation
to allow for particles to be optional, and to generalize "obstacle"
to any list of entities. It also allows for multiple particle
Entities, in case you want e.g. particles with different textures
in the same scene.