Make boid steering override configurable
Adds an option to the flocking simulation to either have steering
be the only acceleration on the boid, or if it is just another
source of acceleration for the boid.

The former will never clip the obstacle's sphere, the latter might,
but only does so rarely. They have slightly different behavior,
but both might be desirable.