Render fish for flocking boids
Fish do not currently face the correct direction or bank.
But this sets up the code to render textured models rather
than ColoredMeshes for the boids.

Right now, the design across when we render Models vs
ColoredMeshes is fragile. I really would like to redesign this,
but that must happen at a later date due to deadlines.