Remove trait bound on get_collided_face_from_list
Use the position to derive the velocity assuming linear velocity, and
just accept a position rather than a struct which requires a position
implementation. Removes the Position and Velocity traits.

This is required for cases we want to use this funciton where we
aren't storing the velocity explicitly, such as in rigid body
simulations. Deriving the velocity inside the function
is simpler and better for such cases.