Improve capacity of Entity instance buffer
Before, every Entity used an instance buffer with MAX_INSTANCES
(5000) capacity. This was really just intended for particles.

Now, users can specify if they want additional capacity beyond
the initial instance used to create the Entity. If None is
provided, the buffer is created to be just large enough to
hold the initial instance data. If some capacity is provided,
it's created with that capacity.

This improves performance due to smaller buffers for each Entity,
and will also allow for more flexibility in specifying larger
capacities.